# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JournalEntrySchema": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/LogLevel"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/LogType"
          }
        },
        "required": [
          "id",
          "message",
          "level",
          "type",
          "timestamp"
        ],
        "title": "JournalEntrySchema",
        "type": "object"
      },
      "JournalResponse": {
        "properties": {
          "journal": {
            "items": {
              "$ref": "#/components/schemas/JournalEntrySchema"
            },
            "title": "Journal",
            "type": "array"
          }
        },
        "required": [
          "journal"
        ],
        "title": "JournalResponse",
        "type": "object"
      },
      "LaunchSchema": {
        "properties": {
          "failed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed At"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_aborted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Aborted"
          },
          "scheduled_at": {
            "format": "date-time",
            "title": "Scheduled At",
            "type": "string"
          },
          "scheduled_by": {
            "title": "Scheduled By",
            "type": "string"
          },
          "skipped_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skipped At"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "scheduled_at",
          "scheduled_by"
        ],
        "title": "LaunchSchema",
        "type": "object"
      },
      "LogLevel": {
        "enum": [
          "INFO",
          "WARNING",
          "ERROR"
        ],
        "title": "LogLevel",
        "type": "string"
      },
      "LogType": {
        "enum": [
          "UNCLASSIFIED",
          "FILE"
        ],
        "title": "LogType",
        "type": "string"
      },
      "ScheduleResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSchema"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "tasks"
        ],
        "title": "ScheduleResponse",
        "type": "object"
      },
      "ScopeResponse": {
        "properties": {
          "scope_id": {
            "format": "uuid",
            "title": "Scope Id",
            "type": "string"
          }
        },
        "required": [
          "scope_id"
        ],
        "title": "ScopeResponse",
        "type": "object"
      },
      "ScopedTaskStatus": {
        "enum": [
          "NEW",
          "PENDING",
          "IN_PROGRESS",
          "SUCCESS",
          "FAILED",
          "SKIPPED"
        ],
        "title": "ScopedTaskStatus",
        "type": "string"
      },
      "TaskListResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSchema"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "tasks"
        ],
        "title": "TaskListResponse",
        "type": "object"
      },
      "TaskSchema": {
        "properties": {
          "current_launch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LaunchSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "title": "Depends On",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "latest_launch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LaunchSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "spec_id": {
            "title": "Spec Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ScopedTaskStatus"
          }
        },
        "required": [
          "id",
          "spec_id",
          "label",
          "description",
          "depends_on",
          "status"
        ],
        "title": "TaskSchema",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "DAG-based task runner with state machines and Celery dispatch",
    "title": "Task Orchestrator",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/scopes/{scope_id}": {
      "post": {
        "description": "Create a new job for a scope. Idempotent \u2014 returns 409 if it already exists.",
        "operationId": "init_scope_api_scopes__scope_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Init Scope",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/scopes/{scope_id}/run": {
      "delete": {
        "operationId": "stop_run_api_scopes__scope_id__run_delete",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Stop Run",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/scopes/{scope_id}/tasks": {
      "get": {
        "operationId": "get_tasks_api_scopes__scope_id__tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/scopes/{scope_id}/tasks/{task_id}/launches/{launch_id}": {
      "delete": {
        "operationId": "abort_task_api_scopes__scope_id__tasks__task_id__launches__launch_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "launch_id",
            "required": true,
            "schema": {
              "title": "Launch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Abort Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/scopes/{scope_id}/tasks/{task_id}/launches/{launch_id}/journal": {
      "get": {
        "operationId": "get_journal_api_scopes__scope_id__tasks__task_id__launches__launch_id__journal_get",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "launch_id",
            "required": true,
            "schema": {
              "title": "Launch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Journal",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/scopes/{scope_id}/tasks/{task_id}/schedule": {
      "post": {
        "operationId": "schedule_task_api_scopes__scope_id__tasks__task_id__schedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Schedule Task",
        "tags": [
          "tasks"
        ]
      }
    }
  }
}